Unify the two copies of ParserClient
The one that was previously in almond-dialog-agent and the one
that was previously in tool/lib
