+++ lib/dataset-tools/evaluation/dialogue_evaluator.js
@@ -189,7 +189,7 @@ class DialogueEvaluatorStream extends Stream.Transform {
 
         const targetCode = (await this._target.serializePrediction(goldUserTarget, tokens, entities, 'user')).join(' ');
 
-        const parsed = await this._parser.sendUtterance(tokens.join(' '), true, contextCode, contextEntities);
+        const parsed = await this._parser.sendUtterance(tokens.join(' '), contextCode, contextEntities, { tokenized: true });
 
         const predictions = parsed.candidates
             .filter((beam) => beam.score !== 'Infinity') // ignore exact matches
