+++ lib/dataset-tools/evaluation/sentence_evaluator.js
@@ -179,7 +179,7 @@ class SentenceEvaluator {
             predictions = this._predictions;
         } else {
             try {
-                const parsed = await this._parser.sendUtterance(this._preprocessed, this._tokenized, contextCode, contextEntities);
+                const parsed = await this._parser.sendUtterance(this._preprocessed, contextCode, contextEntities, { tokenized: this._tokenized });
                 if (!entities)
                     entities = parsed.entities;
