+++ lib/dialogue-agent/almond.js
@@ -14,10 +14,11 @@ const events = require('events');
 const interpolate = require('string-interp');
 const ThingTalk = require('thingtalk');
 
+const ParserClient = require('../prediction/parserclient');
+
 const Semantic = require('./semantic');
 const Intent = Semantic.Intent;
 const ValueCategory = Semantic.ValueCategory;
-const ParserClient = require('./parserclient');
 const Dispatcher = require('./dispatcher');
 
 const DEFAULT_GETTEXT = {
@@ -111,8 +112,8 @@ module.exports = class Almond extends events.EventEmitter {
         this._debug = !!this._options.debug;
 
         this._delegate = delegate;
-        this.parser = new ParserClient(this._options.sempreUrl,
-            engine.platform.locale, engine.platform.getSharedPreferences(), engine.platform);
+        this._prefs = engine.platform.getSharedPreferences();
+        this.parser = ParserClient.get(this._options.sempreUrl, engine.platform.locale, engine.platform);
 
         this._dispatcher = new Dispatcher(this, this._engine, this._debug);
         this._choices = [];
@@ -200,8 +201,13 @@ module.exports = class Almond extends events.EventEmitter {
         this._raw = (expecting === ValueCategory.RawString || expecting === ValueCategory.Password);
     }
 
-    start() {
+    async start() {
+        await this._parser.start();
         return this._dispatcher.start(this._options.showWelcome, this._options.configureMessaging);
+    }
+
+    async stop() {
+        await this._parser.stop();
     }
 
     _isUnsupportedError(e) {
@@ -359,6 +365,25 @@ module.exports = class Almond extends events.EventEmitter {
         }
     }
 
+    _sendUtterance(utterance) {
+        let contextCode, contextEntities;
+        if (this._prefs.get('experimental-contextual-model')) {
+            const now = new Date;
+            if (this._context.timeout > now) {
+                contextCode = this._context.code;
+                contextEntities = this._context.entities;
+            } else {
+                contextCode = 'null';
+                contextEntities = {};
+            }
+        }
+        return this._parser.sendUtterance(utterance, contextCode, contextEntities, {
+            expect: this._expecting,
+            choices: this._choices,
+            store: this._prefs.get('sabrina-store-log') || 'no'
+        });
+    }
+
     handleCommand(command, platformData = {}, postprocess) {
         this.stats.hit('sabrina-command');
         this.emit('active');
@@ -374,7 +399,7 @@ module.exports = class Almond extends events.EventEmitter {
                 return this._doHandleCommand(intent, command, [], true);
             }
 
-            const analyzed = await this.parser.sendUtterance(command, this._context, this._expecting, this._choices);
+            const analyzed = await this._sendUtterance(command);
             if (postprocess)
                 postprocess(analyzed);
             return this._continueHandleCommand(command, analyzed, platformData);
