+++ lib/index.js
@@ -18,7 +18,7 @@ const DatasetAugmenter = require('./dataset-tools/augmentation');
 const { DatasetParser, DatasetStringifier } = require('./dataset-tools/parsers');
 const DatasetSplitter = require('./dataset-tools/splitter');
 const { LocalTokenizer, RemoteTokenizer } = require('./tokenizer');
-const Predictor = require('./predictor');
+const Predictor = require('./prediction/predictor');
 const {
     SentenceEvaluatorStream,
     CollectSentenceStatistics,
