+++ lib/prediction/predictor.js
@@ -13,7 +13,7 @@ const assert = require('assert');
 const events = require('events');
 const child_process = require('child_process');
 
-const JsonDatagramSocket = require('./utils/json_datagram_socket');
+const JsonDatagramSocket = require('../utils/json_datagram_socket');
 
 const DEFAULT_QUESTION = 'translate from english to thingtalk';
