+++ lib/prediction/remoteparserclient.js
@@ -1,8 +1,8 @@
 // -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
 //
-// This file is part of Almond
+// This file is part of Genie
 //
-// Copyright 2016 The Board of Trustees of the Leland Stanford Junior University
+// Copyright 2016-2020 The Board of Trustees of the Leland Stanford Junior University
 //
 // Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
 //
@@ -13,57 +13,53 @@ const ThingTalk = require('thingtalk');
 const Tp = require('thingpedia');
 const qs = require('qs');
 
-const URL = 'https://almond-nl.stanford.edu';
-
-module.exports = class ParserClient {
-    constructor(baseUrl, locale, prefs, platform) {
-        this._locale = locale || 'en_US';
-        this._baseUrl = (baseUrl || URL) + '/' + this._locale;
-        this._prefs = prefs;
+module.exports = class RemoteParserClient {
+    constructor(baseUrl, locale, platform) {
+        this._baseUrl = baseUrl + '/' + this._locale;
         this._platform = platform;
     }
 
-    onlineLearn(utterance, code, store = 'automatic') {
-        const data = qs.stringify({
-            q: utterance,
-            target: code.join(' '),
-            store: store,
-            thingtalk_version: ThingTalk.version,
-        });
-        if (this._platform)
-            data.developer_key = this._platform.getDeveloperKey();
-
-        Tp.Helpers.Http.post(this._baseUrl + '/learn', data, { dataContentType: 'application/x-www-form-urlencoded' }).then(() => {
-            console.log(`Sent "${utterance}" to Almond-NNParser for learning`);
-        }).catch((e) => {
-            console.error(`Failed to send "${utterance}" to Almond-NNParser for learning: ${e.message}`);
-        });
+    async start() {
+    }
+    async stop() {
     }
 
-    async sendUtterance(utterance, context, expecting, choices) {
-        const store = this._prefs.get('sabrina-store-log') || 'no';
+    onlineLearn(utterance, code, store = 'automatic') {
         const data = {
             q: utterance,
+            target: code.join(' '),
             store: store,
             thingtalk_version: ThingTalk.version,
         };
         if (this._platform)
             data.developer_key = this._platform.getDeveloperKey();
-        if (expecting)
-            data.expect = String(expecting);
-        if (choices)
-            data.choices = choices.map((c) => c.title);
 
-        if (this._prefs.get('experimental-contextual-model')) {
-            const now = new Date;
-            if (context.timeout > now) {
-                data.context = context.code;
-                data.entities = context.entities;
-            } else {
-                data.context = 'null';
-                data.entities = {};
-            }
+        Tp.Helpers.Http.post(this._baseUrl + '/learn', qs.stringify(data), { dataContentType: 'application/x-www-form-urlencoded' }).then(() => {
+            console.log('Sent "' + utterance + '" to Almond-NNParser for learning');
+        }).catch((e) => {
+            console.error('Failed to send "' + utterance + '" to Almond-NNParser for learning: ' + e.message);
+        });
+    }
+
+    async sendUtterance(utterance, contextCode, contextEntities, options = {}) {
+        const data = {
+            q: utterance,
+            store: options.store || 'no',
+            thingtalk_version: ThingTalk.version,
+            tokenized: options.tokenized,
+            skip_typechecking: options.skip_typechecking
+        };
+        if (contextCode !== undefined) {
+            data.context = contextCode.join(' ');
+            data.entities = contextEntities;
         }
+        if (this._platform)
+            data.developer_key = this._platform.getDeveloperKey();
+        if (options.expect)
+            data.expect = String(options.expect);
+        if (options.choices)
+            data.choices = options.choices.map((c) => c.title);
+
         const response = await Tp.Helpers.Http.post(`${this._baseUrl}/query`, JSON.stringify(data), {
             dataContentType: 'application/json' //'
         });
@@ -76,10 +72,10 @@ module.exports = class ParserClient {
         return parsed;
     }
 
-    async generateUtterance(context, targetAct) {
+    async generateUtterance(contextCode, contextEntities, targetAct) {
         const data = {
-            context: context.code,
-            entities: context.entities,
+            context: contextCode,
+            entities: contextEntities,
             target: targetAct
         };
         const response = await Tp.Helpers.Http.post(`${this._baseUrl}/answer`, JSON.stringify(data), {
