+++ tool/predict.js
@@ -15,7 +15,7 @@ const Stream = require('stream');
 const { DatasetParser, DatasetStringifier } = require('../lib/dataset-tools/parsers');
 const StreamUtils = require('../lib/utils/stream-utils');
 const { maybeCreateReadStream, readAllLines } = require('./lib/argutils');
-const ParserClient = require('./lib/parserclient');
+const ParserClient = require('../lib/prediction/parserclient');
 
 class PredictStream extends Stream.Transform {
     constructor(parser, tokenized, debug) {
@@ -27,7 +27,7 @@ class PredictStream extends Stream.Transform {
     }
     
     async _process(ex) {
-        const parsed = await this._parser.sendUtterance(ex.preprocessed, this._tokenized, ex.context);
+        const parsed = await this._parser.sendUtterance(ex.preprocessed, ex.context, {}, { tokenized: this._tokenized });
 
         const predictions = parsed.candidates
             .filter((beam) => beam.score !== 'Infinity') // ignore exact matches
